kubernetes: add ingress controller
An ingress controller is the first step in exposing our internal
kubernetes services to the outside world. We've decided on using the
nginx ingress controller because it is endorsed by kubernetes project,
ensuring longevity, and nginx is something many of us at the OCF are
already familiar with.

Right now there doesn't appear to be an established way of getting
external traffic into a kubernetes cluster _on bare metal_. The solution
here is to set up an ingress controller as a NodePort service on every
kubernetes worker (note: for future-proofing we can run this only on a
subset of our workers) that will redirect traffic to the correct worker
based on the "Host" header.

One of the main arguments _against_ using NodePort is that there are a
limited number of ports for use on a system. Since all of our traffic
right now is http(s), we will only need to reserve two ports.  In the
future, if we start running other TCP and UDP services in the cluster
that need to be exposed, we may have to introduce something more
complex, like MetalLB.