kubernetes: add loadbalancer
This commit is closely related to 24aa63a, in which ingress was added to
kubernetes. Since all of the kubernetes worker nodes have an exposed
ingress controller, we bolt HAProxy on the master to more evenly
distribute traffic. This also gives us the added benefit of routing all
external traffic to kubernetes through one external IP.

When we add multiple masters we will use keepalived to ensure the load
balancer continues to listens on one external IP.